Tweak the UI

+++ app/src/main/java/com/repkap11/repweather/api/rest/Clouds.java
@@ -13,6 +13,6 @@ public class Clouds {
      */
     @SerializedName("all")
     @Expose
-    public Float all;
+    public Integer all;
 
 }
